requirements.txt
numpy==1.21.2
Pillow==9.2.0
torch==1.9.0
transformers==4.22.1
opencv-python==4.5.1.48
tqdm==4.64.0
matplotlib==3.4.3